requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Core Dependencies
PyMuPDF==1.23.8
opencv-python==4.9.0.80
numpy==1.26.3
python-docx==1.1.0
pytesseract==0.3.10
ocrmypdf==16.0.0
tqdm==4.66.1

# AI Model
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99

# Utilities
pydantic==2.5.3
aiofiles==23.2.1
